Move relevant types to schema
