+++ examples/pos-mainchain/src/app/modules/testNft/schema.ts
@@ -1,0 +1,79 @@
+/*
+ * Copyright © 2023 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
+
+import {
+	LENGTH_COLLECTION_ID,
+	LENGTH_NFT_ID,
+	MAX_LENGTH_MODULE_NAME,
+	MIN_LENGTH_MODULE_NAME,
+} from './constants';
+
+export const mintNftParamsSchema = {
+	$id: '/lisk/nftTransferParams',
+	type: 'object',
+	required: ['address', 'collectionID', 'attributesArray'],
+	properties: {
+		address: {
+			dataType: 'bytes',
+			format: 'lisk32',
+			fieldNumber: 1,
+		},
+		collectionID: {
+			dataType: 'bytes',
+			minLength: LENGTH_COLLECTION_ID,
+			maxLength: LENGTH_COLLECTION_ID,
+			fieldNumber: 2,
+		},
+		attributesArray: {
+			type: 'array',
+			fieldNumber: 4,
+			items: {
+				type: 'object',
+				required: ['module', 'attributes'],
+				properties: {
+					module: {
+						dataType: 'string',
+						minLength: MIN_LENGTH_MODULE_NAME,
+						maxLength: MAX_LENGTH_MODULE_NAME,
+						pattern: '^[a-zA-Z0-9]*$',
+						fieldNumber: 1,
+					},
+					attributes: {
+						dataType: 'bytes',
+						fieldNumber: 2,
+					},
+				},
+			},
+		},
+	},
+};
+
+export const destroyNftParamsSchema = {
+	$id: '/lisk/nftDestroyParams',
+	type: 'object',
+	required: ['address', 'nftID'],
+	properties: {
+		address: {
+			dataType: 'bytes',
+			format: 'lisk32',
+			fieldNumber: 1,
+		},
+		nftID: {
+			dataType: 'bytes',
+			minLength: LENGTH_NFT_ID,
+			maxLength: LENGTH_NFT_ID,
+			fieldNumber: 2,
+		},
+	},
+};
